Separate module for maps and atlas data exporters
