+++ PyPoE/cli/exporter/wiki/parsers/lua.py
@@ -185,29 +185,6 @@ class LuaHandler(ExporterHandler):
         lua_sub = self.parser.add_subparsers()
 
         parser = lua_sub.add_parser(
-            "atlas",
-            help="Extract Atlas information not covered by maps",
-        )
-        self.add_default_parsers(
-            parser=parser,
-            cls=AtlasParser,
-            func=AtlasParser.main,
-        )
-        group = parser.add_mutually_exclusive_group(required=False)
-        group.add_argument(
-            "-sn",
-            "--series-name",
-            help="Export Atlas nodes data for map series by name (localized)",
-            dest="series_name",
-        )
-        group.add_argument(
-            "-sid",
-            "--series-id",
-            help="Export Atlas nodes data for map series by internal ID",
-            dest="series_id",
-        )
-
-        parser = lua_sub.add_parser(
             "bestiary",
             help="Extract bestiary information",
         )
@@ -326,110 +303,6 @@ class LuaHandler(ExporterHandler):
             cls=MinimapIconsParser,
             func=MinimapIconsParser.main,
         )
-
-
-class AtlasParser(GenericLuaParser):
-    _files = ["MapSeries.datc64"]
-
-    def has_tier_data(self, series_id):
-        return f"{series_id}Tier" in self.rr["MapSeriesTiers.dat64"].specification.columns_all
-
-    def main(self, parsed_args):
-
-        r = ExporterResult()
-        series = None
-
-        if parsed_args.series_id:
-            if not self.has_tier_data(parsed_args.series_id):
-                # should we bail out here or is there anything to export for Atlas of Worlds and before?
-                raise Exception(f"Tier data not available for {parsed_args.series_id}")
-            self.rr["MapSeries.dat64"].build_index("Id")
-            series = self.rr["MapSeries.dat64"].index["Id"][parsed_args.series_id]
-        elif parsed_args.series_name:
-            self.rr["MapSeries.dat64"].build_index("Name")
-            series_rows = self.rr["MapSeries.dat64"].index["Name"][parsed_args.series_name]
-            if not series_rows:
-                raise Exception(f"No tier found with name {parsed_args.series_name}")
-            if not self.has_tier_data(series_rows[0]["Id"]):
-                raise Exception(f"Tier data not available for {parsed_args.series_name}")
-            series = series_rows[0]
-        else:
-            series = self.rr["MapSeries.dat64"][-1]
-
-        self.rr["AtlasNode.dat64"].build_index("WorldAreasKey")
-
-        output = []
-
-        if series:
-            series_id = series["Id"]
-            for tier in self.rr["MapSeriesTiers.dat64"]:
-                tier_number = tier[f"{series_id}Tier"]
-                if tier_number == 0:
-                    continue
-
-                map_base = tier["MapsKey"]
-
-                for world_area in (
-                    map_base["Regular_WorldAreasKey"],
-                    map_base["Unique_WorldAreasKey"],
-                ):
-                    if not world_area:
-                        continue
-                    node_data = {
-                        "world_area": world_area["Id"],
-                        "base_item": map_base["BaseItemTypesKey"]["Id"],
-                        "tier_0": tier_number,
-                    }
-                    output.append(node_data)
-
-                    if series != self.rr["MapSeries.dat64"][-1]:
-                        # AtlasNode.dat only contains data for current series
-                        continue
-
-                    atlas_node = self.rr["AtlasNode.dat64"].index["WorldAreasKey"][world_area]
-                    if atlas_node:
-                        atlas_node = atlas_node[0]
-                        for n in range(5):
-                            node_data[f"tier_{n}"] = atlas_node[f"Tier{n}"]
-                        node_data["connections"] = [
-                            conn["WorldAreasKey"]["Id"] for conn in atlas_node["AtlasNodeKeys"]
-                        ]
-                        node_data["flavour_text"] = atlas_node["FlavourTextKey"]["Text"]
-                        node_data["not_on_atlas"] = atlas_node["NotOnAtlas"]
-                        node_data["div_cards"] = [card["Id"] for card in atlas_node["DivCards"]]
-
-            r.add_result(
-                text=LuaFormatter.format_module(sorted(output, key=lambda x: x["world_area"])),
-                out_file="atlas_nodes_%s.lua" % series_id,
-                wiki_page=[
-                    {
-                        "page": "Module:Atlas/nodes_%s" % series_id,
-                        "condition": None,
-                    }
-                ],
-            )
-
-        r.add_result(
-            text=LuaFormatter.format_module(
-                [
-                    {
-                        "ordinal": i,
-                        "id": tier["Id"],
-                        "name": tier["Name"]
-                    }
-                    for i, tier in enumerate(self.rr["MapSeries.dat64"], 1)
-                ]
-            ),
-            out_file="map_series.lua",
-            wiki_page=[
-                {
-                    "page": "Module:Atlas/map series",
-                    "condition": None,
-                }
-            ],
-        )
-
-        return r
 
 
 class MinimapIconsParser(GenericLuaParser):
